// src/app/api/model/fournisseurDto.ts
/**
 * OpenApi specification - Ousmane Vincent Dione
 * OpenApi documentation for Stock Management
 *
 * The version of the OpenAPI document: 1.0
 * Contact: dioneousmanevincent@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdresseDto } from './adresseDto';
import { CommandeFournisseurDto } from './commandeFournisseurDto';


export interface FournisseurDto { 
    id?: number;
    nom?: string;
    prenom?: string;
    adresseDto?: AdresseDto;
    photo?: string;
    email?: string;
    numTel?: string;
    commandeFournisseurDtos?: Array<CommandeFournisseurDto>;
}